Extract decode different types
